@@ -2,27 +2,27 @@
     "name": "vantiv/payment",
     "description": "Vantiv Payment Extension",
     "require": {
-        "php": "7.0.2|7.0.4|~7.0.6|~7.1.0",
-        "magento/framework": "101.0.*",
-        "magento/module-config": "101.0.*",
-        "magento/module-directory": "100.2.*",
-        "magento/module-payment": "100.2.*",
-        "magento/module-checkout": "100.2.*",
-        "magento/module-sales": "101.0.*",
-        "magento/module-backend": "100.2.*",
-        "magento/module-vault": "101.0.*",
-        "magento/module-customer": "101.0.*",
-        "magento/module-catalog": "102.0.*",
-        "magento/module-quote": "101.0.*",
-        "magento/module-paypal": "100.2.*",
-        "magento/module-theme": "100.2.*",
-        "magento/module-ui": "101.0.*"
+        "php": "~7.2.0|~7.3.0",
+        "magento/framework": "*",
+        "magento/module-config": "*",
+        "magento/module-directory": "*",
+        "magento/module-payment": "*",
+        "magento/module-checkout": "*",
+        "magento/module-sales": "*",
+        "magento/module-backend": "*",
+        "magento/module-vault": "*",
+        "magento/module-customer": "*",
+        "magento/module-catalog": "*",
+        "magento/module-quote": "*",
+        "magento/module-paypal": "*",
+        "magento/module-theme": "*",
+        "magento/module-ui": "*"
     },
     "suggest": {
-        "magento/module-checkout-agreements": "100.2.*"
+        "magento/module-checkout-agreements": "*"
     },
     "type": "magento2-module",
-    "version": "100.2.0",
+    "version": "*",
     "license": [
         "proprietary"
     ],
